removed a comment about it being an in-memory database, it is not anymore. Also added why the test cannot run on an in-memory database with the current code

+++ tests/test_storm_db_manager.py
@@ -14,12 +14,15 @@ class TestStormDBManager(TestCase):
 
     def setUp(self):
         super(TestStormDBManager, self).setUp()
+        # Do not use an in-memory database. Different connections to the same
+        # in-memory database do not point towards the same database.
+        # See: http://blog.devork.be/2010/11/storm-and-sqlite-in-memory-databases.html
         self.storm_db = StormDBManager("sqlite:%s" % self.SQLITE_TEST_DB)
-
     def tearDown(self):
         super(TestStormDBManager, self).tearDown()
-        # Delete the database file.
-        os.unlink(self.SQLITE_TEST_DB)
+        # Delete the database file if not using an in-memory database.
+        if os.path.exists(self.SQLITE_TEST_DB):
+            os.unlink(self.SQLITE_TEST_DB)
 
     def create_car_database(self):
         """
